LPD-44969 Move util class to API
